Updated the documentation. Style clean-up.
